@@ -6,14 +6,14 @@
         "php": "~8.1.0 || ~8.2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^10.5",
-        "symfony/var-dumper": "^5.2",
-        "phpstan/phpstan": "^1.9",
-        "phpstan/phpstan-phpunit": "^1.2",
-        "illuminate/database": "^9.43",
-        "doctrine/dbal": "^3.5",
-        "infection/infection": "^0.26.16",
-        "friendsofphp/php-cs-fixer": "^3.14"
+        "phpunit/phpunit": "^10.5.12",
+        "symfony/var-dumper": "^5.4.36",
+        "phpstan/phpstan": "^1.10.60",
+        "phpstan/phpstan-phpunit": "^1.3.16",
+        "illuminate/database": "^9.52.16",
+        "doctrine/dbal": "^3.8.3",
+        "infection/infection": "^0.26.21",
+        "friendsofphp/php-cs-fixer": "^3.51"
     },
     "autoload": {
         "psr-4": {
